feat(primitives): restructure integer as integer64 with alias
Move signed integer to integer/integer64 with canonical int64 and alias int. Add Registry::register_type_alias with comparison activation and align file naming.

+++ crates/primitives/src/integer/integer64/literal.rs
@@ -7,10 +7,11 @@ pub(crate) fn parse(raw_text: &str, type_id: TypeId) -> Result<Value, CoreError>
         .map(|value| Value::new(type_id, value))
         .map_err(|error| CoreError::InvalidLiteral {
             raw_text: raw_text.into(),
-            type_name: "int".into(),
+            type_name: "int64".into(),
             message: error.to_string(),
         })
 }
+
 #[cfg(test)]
 #[path = "literal.tests.rs"]
 mod tests;
